w3c-reports/12092014/mac/runned-tests/webaudio/the-audio-api/the-waveshapernode-interface/curve-tests-expected.txt

CONSOLE MESSAGE: line 185: TypeError: Not enough arguments

Harness Error. harness_status.status = 1 , harness_status.message = TypeError: Not enough arguments
